Move user management logic to its own module
