+++ server.js
@@ -6,28 +6,8 @@ const server = http.createServer(app);
 const io = new Server(server, { cors: true });
 
 const PORT = process.env.PORT || 7890;
-const ROOM_THRESHOLD = 3;
 
-const users = {};
-let roomNumber = 0;
-let unassignedUsers = 0;
-
-const addUser = socket => {
-  console.log(`new user in room ${roomNumber}: ${socket.id}`);
-  unassignedUsers += 1;
-  users[socket.id] = socket;
-  users[socket.id].room = String(roomNumber);
-
-  if(unassignedUsers >= ROOM_THRESHOLD) {
-    roomNumber ++;
-    unassignedUsers = 0;
-  }
-};
-
-const deleteUser = socket => {
-  console.log('user left: ', socket.id);
-  delete users[socket.id];
-};
+const { users, addUser, deleteUser } = require('./lib/socket/user-utils');
 
 const onConnection = socket => {
   addUser(socket);
